media: v4l: subdev: Add a helper to ensure locking of active state
Add a helper to lock the active state when a state (try or active) is
already locked, but it's not known which one it is (and it does not
matter). Effectively this is useful in cases where there is a dependency
to active state also in try states and access to that needs to be
serialised.

Note that the try state lock must always be acquired before the active
state lock. Normally there is no issue with this as the framework locks
the try state first, always.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>